Fix #4509 and add regression test
