+++ nexus/src/app/sagas/mod.rs
@@ -36,6 +36,7 @@ pub mod snapshot_create;
 pub mod snapshot_delete;
 pub mod switch_port_settings_apply;
 pub mod switch_port_settings_clear;
+pub mod switch_port_settings_common;
 pub mod test_saga;
 pub mod volume_delete;
 pub mod volume_remove_rop;
